Meal - stats by day and by week

+++ nutrition-app/src/main/java/org/nutrition/app/food/dto/NutritionMineralsDTO.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 
 public record NutritionMineralsDTO (
         UUID id,
-        float portionSize,
+        Double portionSize,
         String unit,
         Double calcium,
         Double iron,
